# agent/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PG0 agent API",
    "version": "1.0.0",
    "description": "Write, syntax-check, run and store programs written in PG0 / PG0.5. The full manual (language specification and library reference included) is served as Markdown at GET /manual."
  },
  "servers": [{"url": "/agent/v1"}],
  "components": {
    "securitySchemes": {
      "bearer": {"type": "http", "scheme": "bearer", "description": "Only required when the server is configured with an API key."}
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {"type": "string", "example": "invalid_request"},
              "message": {"type": "string"}
            }
          }
        }
      },
      "RunRequest": {
        "type": "object",
        "required": ["code"],
        "properties": {
          "code": {"type": "string", "description": "Program source code (UTF-8). Lines are separated by \\n."},
          "input": {
            "description": "Lines returned by input(), one per call. A string is split at newlines; an array is used as is.",
            "oneOf": [{"type": "string"}, {"type": "array", "items": {"type": "string"}}]
          },
          "mode": {"type": "string", "enum": ["PG0", "PG0.5"], "default": "PG0.5", "description": "Language mode. PG0 = integers only, no functions. PG0.5 = full language."},
          "lang": {"type": "string", "enum": ["en", "ja"], "default": "en", "description": "Language of error messages."},
          "timeout_ms": {"type": "integer", "default": 5000, "description": "Execution time limit in milliseconds (capped by the server)."},
          "max_steps": {"type": "integer", "default": 10000000, "description": "Maximum number of executed statements (capped by the server)."},
          "variables": {"type": "boolean", "default": true, "description": "Include the final values of global variables in the response."}
        }
      },
      "RunResult": {
        "type": "object",
        "properties": {
          "status": {"type": "string", "enum": ["ok", "error", "timeout", "step_limit", "memory_limit"]},
          "mode": {"type": "string", "enum": ["PG0", "PG0.5"]},
          "output": {"type": "string", "description": "Everything written by print() and println()."},
          "output_truncated": {"type": "boolean"},
          "error_output": {"type": "string", "description": "Everything written by error(), one line per call."},
          "result": {"description": "Value passed to exit, converted to JSON (null when none). Arrays with keys become objects.", "nullable": true},
          "result_type": {"type": "string", "enum": ["integer", "float", "string", "array"], "nullable": true},
          "error": {
            "type": "object",
            "nullable": true,
            "properties": {
              "message": {"type": "string"},
              "line": {"type": "integer", "nullable": true, "description": "1-based line number in the program."},
              "source": {"type": "string", "nullable": true, "description": "Source text of that line."},
              "phase": {"type": "string", "enum": ["parse", "runtime", "internal"]}
            }
          },
          "variables": {"type": "object", "nullable": true, "additionalProperties": true, "description": "Final values of global variables, converted to JSON."},
          "stats": {
            "type": "object",
            "properties": {
              "steps": {"type": "integer", "nullable": true},
              "elapsed_ms": {"type": "integer"},
              "input_lines_used": {"type": "integer", "nullable": true}
            }
          }
        }
      },
      "CheckRequest": {
        "type": "object",
        "required": ["code"],
        "properties": {
          "code": {"type": "string"},
          "mode": {"type": "string", "enum": ["PG0", "PG0.5"], "default": "PG0.5"},
          "lang": {"type": "string", "enum": ["en", "ja"], "default": "en"}
        }
      },
      "CheckResult": {
        "type": "object",
        "properties": {
          "ok": {"type": "boolean"},
          "mode": {"type": "string"},
          "error": {"nullable": true, "type": "object", "properties": {"message": {"type": "string"}, "line": {"type": "integer", "nullable": true}, "source": {"type": "string", "nullable": true}, "phase": {"type": "string"}}}
        }
      },
      "ScriptSummary": {
        "type": "object",
        "properties": {
          "cid": {"type": "string", "description": "Script id (UUID)."},
          "name": {"type": "string"},
          "author": {"type": "string"},
          "mode": {"type": "string", "enum": ["PG0", "PG0.5"]},
          "private": {"type": "integer", "enum": [0, 1]},
          "createTime": {"type": "integer", "nullable": true, "description": "Milliseconds since 1970-01-01 UTC."},
          "updateTime": {"type": "integer"},
          "url": {"type": "string", "description": "URL that opens the script in the web editor."}
        }
      },
      "Script": {
        "allOf": [
          {"$ref": "#/components/schemas/ScriptSummary"},
          {"type": "object", "properties": {"memo": {"type": "string"}, "code": {"type": "string"}}}
        ]
      },
      "ScriptCreate": {
        "type": "object",
        "required": ["name", "code", "password"],
        "properties": {
          "name": {"type": "string", "maxLength": 100, "description": "Script name. Must be unique among public scripts."},
          "code": {"type": "string"},
          "password": {"type": "string", "description": "Needed later to update or delete the script. The same password works in the web editor."},
          "author": {"type": "string", "maxLength": 100, "default": "AI agent"},
          "memo": {"type": "string", "description": "Free text description."},
          "private": {"type": "integer", "enum": [0, 1], "default": 0, "description": "1 hides the script from public lists (still reachable by cid)."},
          "mode": {"type": "string", "enum": ["PG0", "PG0.5"], "default": "PG0.5"},
          "uuid": {"type": "string", "description": "Optional owner id used to group scripts in listings."}
        }
      },
      "ScriptUpdate": {
        "type": "object",
        "required": ["password"],
        "properties": {
          "password": {"type": "string"},
          "name": {"type": "string"},
          "code": {"type": "string"},
          "author": {"type": "string"},
          "memo": {"type": "string"},
          "private": {"type": "integer", "enum": [0, 1]},
          "mode": {"type": "string", "enum": ["PG0", "PG0.5"]},
          "uuid": {"type": "string"}
        }
      },
      "ScriptRunRequest": {
        "type": "object",
        "properties": {
          "input": {"oneOf": [{"type": "string"}, {"type": "array", "items": {"type": "string"}}]},
          "mode": {"type": "string", "enum": ["PG0", "PG0.5"], "description": "Defaults to the stored script's mode."},
          "lang": {"type": "string", "enum": ["en", "ja"]},
          "timeout_ms": {"type": "integer"},
          "max_steps": {"type": "integer"},
          "variables": {"type": "boolean"}
        }
      }
    }
  },
  "security": [{"bearer": []}, {}],
  "paths": {
    "/": {
      "get": {"summary": "API index: endpoints, manual URLs and limits", "responses": {"200": {"description": "Index"}}}
    },
    "/manual": {
      "get": {
        "summary": "Full manual as Markdown (API usage, language specification, library reference)",
        "parameters": [{"name": "lang", "in": "query", "schema": {"type": "string", "enum": ["en", "ja"]}, "description": "Defaults to Accept-Language, then English."}],
        "responses": {"200": {"description": "Markdown text", "content": {"text/markdown": {"schema": {"type": "string"}}}}}
      }
    },
    "/openapi.json": {
      "get": {"summary": "This document", "responses": {"200": {"description": "OpenAPI document"}}}
    },
    "/libraries": {
      "get": {"summary": "Machine readable list of built-in functions and libraries", "responses": {"200": {"description": "Library list"}}}
    },
    "/check": {
      "post": {
        "summary": "Syntax check a program without running it",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CheckRequest"}}}},
        "responses": {
          "200": {"description": "Check result", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CheckResult"}}}},
          "400": {"description": "Invalid request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}
        }
      }
    },
    "/run": {
      "post": {
        "summary": "Run a program",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RunRequest"}}}},
        "responses": {
          "200": {"description": "Run result (also for programs that fail; see status)", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RunResult"}}}},
          "400": {"description": "Invalid request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
          "413": {"description": "Code or input too large", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
          "429": {"description": "Too many concurrent runs; retry shortly", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}
        }
      }
    },
    "/scripts": {
      "get": {
        "summary": "List or search stored scripts",
        "parameters": [
          {"name": "q", "in": "query", "schema": {"type": "string"}, "description": "Space separated words matched against name and author."},
          {"name": "uuid", "in": "query", "schema": {"type": "string"}, "description": "Owner id; scripts with this owner are listed first, private ones included."},
          {"name": "skip", "in": "query", "schema": {"type": "integer", "default": 0}},
          {"name": "count", "in": "query", "schema": {"type": "integer", "default": 30, "maximum": 100}}
        ],
        "responses": {"200": {"description": "Script list", "content": {"application/json": {"schema": {"type": "object", "properties": {"scripts": {"type": "array", "items": {"$ref": "#/components/schemas/ScriptSummary"}}, "skip": {"type": "integer"}, "count": {"type": "integer"}}}}}}}
      },
      "post": {
        "summary": "Store a new script",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ScriptCreate"}}}},
        "responses": {
          "201": {"description": "Created script", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Script"}}}},
          "400": {"description": "Invalid request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
          "409": {"description": "A public script with the same name exists", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}
        }
      }
    },
    "/scripts/{cid}": {
      "parameters": [{"name": "cid", "in": "path", "required": true, "schema": {"type": "string"}}],
      "get": {
        "summary": "Get a stored script including its code",
        "responses": {"200": {"description": "Script", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Script"}}}}, "404": {"description": "Not found"}}
      },
      "put": {
        "summary": "Update a stored script (only the fields given are changed)",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ScriptUpdate"}}}},
        "responses": {"200": {"description": "Updated script", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Script"}}}}, "401": {"description": "Wrong password"}, "404": {"description": "Not found"}, "409": {"description": "Name conflict"}}
      },
      "delete": {
        "summary": "Delete a stored script",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "required": ["password"], "properties": {"password": {"type": "string"}}}}}},
        "responses": {"200": {"description": "Deleted"}, "401": {"description": "Wrong password"}, "404": {"description": "Not found"}}
      }
    },
    "/scripts/{cid}/run": {
      "parameters": [{"name": "cid", "in": "path", "required": true, "schema": {"type": "string"}}],
      "post": {
        "summary": "Run a stored script",
        "requestBody": {"required": false, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ScriptRunRequest"}}}},
        "responses": {"200": {"description": "Run result", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RunResult"}}}}, "404": {"description": "Not found"}}
      }
    },
    "/scripts/{cid}/history": {
      "parameters": [{"name": "cid", "in": "path", "required": true, "schema": {"type": "string"}}],
      "get": {
        "summary": "List previous versions of a stored script (newest first)",
        "parameters": [{"name": "skip", "in": "query", "schema": {"type": "integer"}}, {"name": "count", "in": "query", "schema": {"type": "integer"}}],
        "responses": {"200": {"description": "History list"}}
      }
    },
    "/scripts/{cid}/history/{time}": {
      "parameters": [{"name": "cid", "in": "path", "required": true, "schema": {"type": "string"}}, {"name": "time", "in": "path", "required": true, "schema": {"type": "integer"}, "description": "updateTime of the version"}],
      "get": {
        "summary": "Get a previous version of a stored script",
        "responses": {"200": {"description": "Script version", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Script"}}}}, "404": {"description": "Not found"}}
      }
    }
  }
}